moved vite to subdirectory
