--- deploy.sh
@@ -1,4 +1,0 @@
-sbt frontend/fullLinkJS
-yarn exec vite -- build
-cp dist/index.html dist/200.html
-surge ./dist 'zio-pet-clinic.surge.sh'
