+++ js-frontend/deploy.sh
@@ -1,0 +1,4 @@
+sbt frontend/fullLinkJS
+yarn exec vite -- build
+cp dist/index.html dist/200.html
+surge ./dist 'zio-pet-clinic.surge.sh'
