+++ js-frontend/main.scss
@@ -1,0 +1,7 @@
+@import "tailwindcss/base";
+@import "tailwindcss/components";
+@import "tailwindcss/utilities";
+
+body {
+  background: rgb(245, 245, 244);
+}
