+++ js-frontend/package.json
@@ -1,0 +1,32 @@
+{
+  "name": "example",
+  "version": "0.1.0",
+  "main": "index.js",
+  "license": "MIT",
+  "devDependencies": {
+    "@tailwindcss/forms": "0.5.2",
+    "@tailwindcss/typography": "0.5.2",
+    "postcss": "8.4.14",
+    "postcss-import": "14.1.0",
+    "postcss-preset-env": "7.6.0",
+    "stylelint": "14.8.5",
+    "stylelint-config-recommended": "7.0.0",
+    "stylelint-config-standard": "25.0.0",
+    "tailwindcss": "3.0.24",
+    "vite": "2.9.9",
+    "vite-plugin-html": "^3.2.0"
+  },
+  "browserslist": [
+    "defaults"
+  ],
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview"
+  },
+  "dependencies": {
+    "autoprefixer": "^10.2.5",
+    "highlight.js": "^11.2.0",
+    "sass": "^1.52.1"
+  }
+}
