+++ js-frontend/postcss.config.js
@@ -1,0 +1,10 @@
+module.exports = () => {
+    const plugins = {
+        'postcss-import': {},
+        tailwindcss: {},
+        autoprefixer: {}
+    }
+    return {
+        plugins
+    }
+}
