--- postcss.config.js
@@ -1,10 +1,0 @@
-module.exports = () => {
-    const plugins = {
-        'postcss-import': {},
-        tailwindcss: {},
-        autoprefixer: {}
-    }
-    return {
-        plugins
-    }
-}
